% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/optimization_functions.R
\name{msnmkb}
\alias{msnmkb}
\title{Multistart nmkb}
\usage{
msnmkb(objective, nb_iter = 100, lower, upper, cl = NULL, ...)
}
\arguments{
\item{objective}{function to be minimized}

\item{nb_iter}{number of iterations}

\item{lower}{lower bounds}

\item{upper}{upper bounds}

\item{cl}{number of cores to be used}

\item{...}{other arguments to be passed to nmkb}
}
\value{
list with the best solution
}
\description{
Multistart nmkb
}
