% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict.R
\name{predict.matern52}
\alias{predict.matern52}
\title{Predict Method for Matérn 5/2 Models}
\usage{
\method{predict}{matern52}(object, newdata, ci = NULL, ...)
}
\arguments{
\item{object}{Fitted model object}

\item{newdata}{New data for prediction}

\item{ci}{Confidence interval (optional)}

\item{...}{Additional arguments}
}
\value{
Predictions for new data
}
\description{
Predict Method for Matérn 5/2 Models
}
