% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict.R
\name{predict.rvfl}
\alias{predict.rvfl}
\title{Predict Method for RVFL Models}
\usage{
\method{predict}{rvfl}(object, newdata, ci = NULL, graph = FALSE, ...)
}
\arguments{
\item{object}{Fitted model object}

\item{newdata}{New data for prediction}

\item{ci}{Confidence interval (optional)}

\item{graph}{Whether to create a graph (optional)}

\item{...}{Additional arguments}
}
\value{
Predictions for new data
}
\description{
Predict Method for RVFL Models
}
