% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/update.R
\name{update_params}
\alias{update_params}
\title{Update function}
\usage{
update_params(
  fit_obj,
  newx,
  newy,
  re_clust = TRUE,
  method = c("direct", "polyak"),
  alpha = 0.5
)
}
\arguments{
\item{alpha}{}
}
\description{
Update function
}
