% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bayesopt.R
\name{bayes_opt}
\alias{bayes_opt}
\title{Title}
\usage{
bayes_opt(
  objective,
  lower,
  upper,
  type_acq = c("ei", "ucb"),
  nb_init = 10,
  nb_iter = 25,
  kappa = 1.96,
  method = c("standard", "direct_online", "polyak_online"),
  surrogate_model = c("rvfl", "matern52", "rvfl_emcee", "rf"),
  optim_surr = c("GCV", "loglik", "cv"),
  activation_function = c("relu", "tanh", "sigmoid"),
  type_optim = c("nlminb", "DEoptim", "msnlminb", "randsearch", "none"),
  early_stopping = FALSE,
  abs_tol = 1e-07,
  rel_tol = 0.001,
  seed = 123,
  verbose = TRUE,
  show_progress = TRUE,
  ...
)
}
\arguments{
\item{...}{}
}
\description{
Title
}
